Remove secrets setup command
