+++ crates/kernelle/Cargo.toml
@@ -18,9 +18,12 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 serde_yaml = "0.9"
 dirs = "6.0"
-reqwest.workspace = true 
+reqwest.workspace = true
 tempfile = "3.0"
-chrono = { version = "0.4", features = ["serde"] } 
+chrono = { version = "0.4", features = ["serde"] }
+secrets = { path = "../secrets" }
+rpassword = "7.0"
+bentley = { path = "../bentley" }
 
 [dev-dependencies]
 tempfile = "3.0"
