+++ crates/kernelle/src/commands/mod.rs
@@ -3,6 +3,7 @@ pub mod daemon;
 pub mod r#do;
 pub mod remove;
 pub mod retrieve;
+pub mod secrets;
 pub mod store;
 pub mod update;
 pub mod version;
